// packages/react-virtuoso/src/interfaces.ts
export interface FlatScrollIntoViewLocation extends ScrollIntoViewLocationOptions {
  /** The index of the item to scroll into view */
  index: number
}

/**
 * @group Common
 */
export type FollowOutput = FollowOutputCallback | FollowOutputScalarType

/**
 * @group Common
 */
export type FollowOutputCallback = (isAtBottom: boolean) => FollowOutputScalarType

/**
 * @group Common
 */
export type FollowOutputScalarType = 'auto' | 'smooth' | boolean
// ...
export interface GroupedScrollIntoViewLocation extends ScrollIntoViewLocationOptions {
  /** The index of the group to scroll into view */
  groupIndex: number
}
// ...
export type ScrollIntoViewLocation = FlatScrollIntoViewLocation | GroupedScrollIntoViewLocation

/**
 * Options for scrolling an item into view.
 * @group Common
 */
export interface ScrollIntoViewLocationOptions {
  /** How to align the item within the viewport */
  align?: 'center' | 'end' | 'start'
  /** The scroll behavior - 'smooth' for animated scrolling, 'auto' for instant */
  behavior?: 'auto' | 'smooth'
  /**
   * Use this function to fine-tune the scrollIntoView behavior.
   * The function receives the item's top and bottom position in the viewport, and the viewport top/bottom.
   * Return an location object to scroll, or null to prevent scrolling.
   * Here's the default implementation:
   * ```ts
const defaultCalculateViewLocation: CalculateViewLocation = ({
  itemTop,
  itemBottom,
  viewportTop,
  viewportBottom,
  locationParams: { behavior, align, ...rest },
}) => {
  if (itemTop < viewportTop) {
    return { ...rest, behavior, align: align ?? 'start' }
  }
  if (itemBottom > viewportBottom) {
    return { ...rest, behavior, align: align ?? 'end' }
  }
  return null
}
   *```
   */
  calculateViewLocation?: CalculateViewLocation
  /**
   * Will be called when the scroll is done, or immediately if no scroll is needed.
   */
  done?: () => void
}